Add register set block with min features
The register block currently can only read and write a fixed
number of bytes. Read/Write interface is done, additionally
there is a output which contains all current values of regs.
This output allows all parts of the sd host that require
some value of the regs to config to recieve it as a
constant input instead of having to request the data
through the read/write cpu interface.